# Rewrite old attribute mapping to save many object allocations