@amatsuda 2017-10-25
@oniofchaos Thanks, this change makes sense.
But can you please rebase from master and `push -f` here instead of merging origin/master into your branch?